refactor: modularize bridge.py and apply style guide improvements
- Extract GooseACPClient to goose_acp_client.py
- Extract MattermostAPI to mattermost_api.py
- Extract configuration to config.py
- Extract utilities to utils.py
- Improve docstrings and type hinting
- Fixes #10
